Change upload of entire context to explicit ADD, update comments.
-Uploading the entire context just to get files in '/data' and '/log' in the
very, very slight chance that there were any was not the way to do it.
So explicitly adding '/config' and '/build-env' was better.
-This moved the job of adding files to these directores to the Hub
Dockerfile itself. It also now allows for remote downloading of data or
source into '/data' by using the Hub Dockerfile's unused ENTRYPOINT
command.
-For some reason, uploading the context also added about 10mb of phantom
data which I could not pin down. That nearly doubled the busyboxplus
image that I worked so hard to make small; couldn't have it.