Update comments and instructions. Mention .dockerignore.
Update lots of comments everywhere.

The actual .dockerignore changes got eaten up by another commit, but the
instructions were updated here. The big change is that now the Hub will
upload the entire context, catching any combination of /config, /data,
folders and other files (build-env) that we put there. This was more
flexible and easier to design for now that we can blacklist items from
the context.

The original reason for not doing this was phantom image size
creep when uploading the entire context. The image seemed to grow by
more then what was in the context. That seems to have disapeared now.