feature: allow no configuration on static/dynamic hub builds
Continuing a policy of warn-but-allow to create a better user
experience. This commit allows for zero configuration uploaded
statically (empty config dirs) and dynamically (via emptry $WHEEL_REPO
variable). It gives a warning but lets the user proceed anyway in case
there are situations where the hub is a placeholder or serves other
purposes then just configuration.